Skybox/PlaneMesh
Added cloudy skybox, it's not the best looking so it'll be changed eventually but that's just a matter of swapping names in the script for the images to render. Added a plane mesh and linked the transform controls to it. The planes MTL isn't loading properly, that's next on the debug list.